move graphql queries inline with components
